Workout screen enhancment (#6)
* refactor: Update routes and components for workouts in App.tsx

* refactor: Update react-select npm dependency to version 5.8.0

* feat: Update muscle group and equipment selection in WorkoutForm

The code changes in `WorkoutForm.tsx` modify the way muscle groups and equipment are selected. The `value` prop for the `StyledSelect` components is updated to use the `map` function instead of optional chaining. This ensures that the selected muscle groups and equipment are correctly displayed in the dropdown.

Based on the recent user commits and repository commits, the suggested commit message is "feat: Update muscle group and equipment selection in WorkoutForm".

* feat: Update muscle group and equipment selection in WorkoutForm

* refactor: Update background color for active navigation tab in Header component

The code changes in `Header.tsx` update the background color for the active navigation tab. The `background-color` property is changed from using the `accent` color to the `navActive` color defined in the theme. This ensures that the active tab stands out and is visually distinguishable from the other tabs.

Based on the recent user commits and repository commits, the suggested commit message is "refactor: Update background color for active navigation tab in Header component".

* refactor: Increase max-width and height for WorkoutForm

The code changes in `WorkoutFormStyles.ts` increase the `max-width` property of the `FormContainer` component to 1200px and the `height` property of the `ImageContainer` component to 500px. This provides a larger cropping experience for images and allows for more content to be displayed within the form.

Based on the recent user commits and repository commits, the suggested commit message is "refactor: Increase max-width and height for WorkoutForm".

* refactor: Increase max-width and height for WorkoutForm

* refactor: Remove unused import statement in App.tsx